Added support for skeleton animation
Added broad support for skeleton animation (initial MD5 import).
Added a different class to deal with the editor (world).
Updated solution targets for Windows build.
Code cleanup.

+++ horseradish/src/common/opengl/tools/immediateMode.cpp
@@ -112,7 +112,7 @@ namespace hr { namespace gl { namespace tools
 		for (size_t i = 0; i < numLeftOvers; i++)
 		{
 			mBufferData[i] = mBufferData[numElements + i];
-			mBufferIndices[i] = i;
+			mBufferIndices[i] = static_cast<unsigned short>(i);
 		}
 
 		mState.curVertex = numLeftOvers;
